requirements.txt
# 基础数据处理
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
pytz>=2024.1

# 数据库
pymysql>=1.1.0

# 网络请求
requests>=2.31.0

# 机器学习
lightgbm>=4.1.0
scikit-learn>=1.3.0

# 其他依赖可以根据需要添加 